refactor(core): move connection_pool (PooledConn/WorkspaceConnectionPool) to julie-core
Copy src/daemon/connection_pool.rs verbatim into crates/julie-core/src/
connection_pool.rs. The original file becomes a 5-line re-export shim so
every crate::daemon::connection_pool::* import path across all consumers
(workspace_pool.rs, database/mod.rs, two test files, workspace/mod.rs)
compiles unchanged — zero call-site edits.

Severs the database→daemon dependency: PooledConn now lives in the leaf
crate; src/database/mod.rs can eventually import from julie_core:: instead
of crate::daemon::.

Gates:
  cargo check                          ✓ (zero errors)
  tests::daemon::connection_pool_test  ✓ 9/9 PASS
  tests::daemon::symbol_db_pooled_test ✓ 5/5 PASS
  cargo build                          ✓ (full workspace)
